Refactoring
Refactor structure of project files.
