+++ lib/bottom_navigation_bar/bottom_navigation_bar.dart
