+++ lib/models/author.dart
