+++ lib/pages/register.dart
@@ -3,12 +3,9 @@ import 'dart:convert';
 import 'package:appmetrica_plugin/appmetrica_plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:jwt_decoder/jwt_decoder.dart';
-import '../author.dart';
-import '../post.dart';
-import '../publication_utils.dart';
-import '../token_storage.dart';
-import '../user_credentials.dart';
-import 'custom_page_route.dart';
+import '../storages/token_storage.dart';
+import '../storages/user_credentials.dart';
+import 'package:green_hub_client/utilities/custom_page_route.dart';
 import 'lenta.dart';
 import 'login.dart';
 import 'package:http/http.dart' as http;
