+++ lib/pages/subscriptions.dart
@@ -3,13 +3,15 @@ import 'dart:convert';
 import 'package:appmetrica_plugin/appmetrica_plugin.dart';
 import 'package:flutter/material.dart';
 import 'package:green_hub_client/pages/profile.dart';
-import 'package:green_hub_client/publication_utils.dart';
+import 'package:green_hub_client/utilities/publication_utils.dart';
+import 'package:green_hub_client/models/author.dart';
+import '../storages/user_credentials.dart';
+import 'package:green_hub_client/bottom_navigation_bar/bottom_navigation_bar.dart';
+import 'package:green_hub_client/bottom_navigation_bar/bottom_navigation_logic.dart';
+import 'package:green_hub_client/utilities/custom_page_route.dart';
+import 'package:green_hub_client/utilities/user_utils.dart';
 
-import '../author.dart';
-import '../user_credentials.dart';
-import 'bottom_navigation_bar.dart';
-import 'bottom_navigation_logic.dart';
-import 'custom_page_route.dart';
+import '../utilities/action_utils.dart';
 
 class Subscriptions extends StatefulWidget {
   @override
@@ -37,7 +39,7 @@ class _SubscriptionsPageState extends State<Subscriptions> with SingleTickerProv
 
     String? currentUsername = UserCredentials().username;
     if (currentUsername != null) {
-      var subscriptionsAndSubscribers = await PublicationUtils.fetchSubscriptionsAndSubscribers(currentUsername);
+      var subscriptionsAndSubscribers = await UserUtils.fetchSubscriptionsAndSubscribers(currentUsername);
       subscriptionList = subscriptionsAndSubscribers[0];
       followerList = subscriptionsAndSubscribers[1];
     }
@@ -117,7 +119,7 @@ class _SubscriptionsListState extends State<SubscriptionsList> {
 
   void _unsubscribeUser(int index) async {
     AppMetrica.reportEvent('Click on "Unsubscribe" button');
-    PublicationUtils.subscribeOrUnsubscribe('https://greenhubapp.ru:80/users/' + widget.users[index].userId.toString() + '/unsubscribe');
+    ActionUtils.subscribeOrUnsubscribe('https://greenhubapp.ru:80/users/' + widget.users[index].userId.toString() + '/unsubscribe');
     // Удалить пользователя из списка
     setState(() {
       widget.users.removeAt(index);
@@ -139,7 +141,7 @@ class _SubscriptionsListState extends State<SubscriptionsList> {
           child: ListTile(
             leading: GestureDetector(
               onTap: () async {
-                Author? author = await PublicationUtils.fetchAuthorByUsername(widget.users[index].username);
+                Author? author = await UserUtils.fetchAuthorByUsername(widget.users[index].username);
                 if (author != null) {
                   Navigator.pushReplacement(
                     context,
@@ -153,7 +155,7 @@ class _SubscriptionsListState extends State<SubscriptionsList> {
             ),
             title: GestureDetector(
               onTap: () async {
-                Author? author = await PublicationUtils.fetchAuthorByUsername(widget.users[index].username);
+                Author? author = await UserUtils.fetchAuthorByUsername(widget.users[index].username);
                 if (author != null) {
                   Navigator.pushReplacement(
                     context,
